+++ index.html
@@ -33,6 +33,6 @@
     </ul>
   </div>
 </section>
-<script type="module" src="js/app.js"></script>
+<script type="module" src="js/TodoApp.js"></script>
 </body>
 </html>
