+++ js/TodoItemFilter.js
@@ -1,0 +1,12 @@
+export class TodoItemFilter {
+
+  constructor() {
+    this.filterCompleted = items => {
+      return items.filter(item => item.completed);
+    }
+
+    this.filterActive = items => {
+      return items.filter(item => !item.completed);
+    }
+  }
+}
